+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.4.0
 - Add Node.js APIs to the list from the global node_modules.
+- Add progress indicator while initializing.
 - Amend list ordering so that node modules are placed first.
 
 ## 1.3.0